pyproject.toml
[project]
name = "api-test"
version = "0.1.0"
description = ""
authors = [
    {name = "RafaelNogueiraXD",email = "rafa10_nogueira@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.128.5,<0.129.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "openai (>=1.107.2,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
packages = [{include = "api_test", from = "src"}]

[tool.taskipy.tasks]
server = "uvicorn api_test.main:app --reload --host 0.0.0.0 --port 8000"
server-prod = "uvicorn api_test.main:app --host 0.0.0.0 --port 8000"
docker-build = "docker compose build"
docker-up = "docker compose up -d"
docker-down = "docker compose down"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
